FormattableContent: Code reorganization
