+++ WordPressKit/FormattableContent.swift
@@ -1,32 +1,5 @@
 import Foundation
 
-public enum ParentKind: String {
-    case Comment        = "comment"
-    case CommentLike    = "comment_like"
-    case Follow         = "follow"
-    case Like           = "like"
-    case Matcher        = "automattcher"
-    case NewPost        = "new_post"
-    case Post           = "post"
-    case User           = "user"
-    case Unknown        = "unknown"
-
-    var toTypeValue: String {
-        return rawValue
-    }
-}
-
-public protocol FormattableContentParent: AnyObject {
-    var metaCommentID: NSNumber? { get }
-    var uniqueID: String? { get }
-    var kind: ParentKind { get }
-    var metaReplyID: NSNumber? { get }
-    var isPingback: Bool { get }
-    func didChangeOverrides()
-    func isEqual(to other: FormattableContentParent) -> Bool
-}
-
-
 // MARK: - FormattableContent Implementation
 //
 public class FormattableContent: Equatable {
@@ -113,8 +86,6 @@ public class FormattableContent: Equatable {
         self.type = nil
     }
 }
-
-
 
 // MARK: - FormattableContent Computed Properties
 //
@@ -208,8 +179,6 @@ extension FormattableContent {
     }
 }
 
-
-
 // MARK: - FormattableContent Methods
 //
 extension FormattableContent {
@@ -291,7 +260,6 @@ extension FormattableContent {
     }
 }
 
-
 // MARK: - FormattableContent Parsers
 //
 extension FormattableContent {
@@ -324,7 +292,6 @@ extension FormattableContent {
         return ranges
     }
 }
-
 
 // MARK: - FormattableContent Types
 //
