+++ WordPressKit/NSMutableAttributedString+extensions.swift
@@ -1,0 +1,18 @@
+import Foundation
+
+extension NSMutableAttributedString {
+    func applyAttributes(toQuotes attributes: [NSAttributedStringKey: Any]?) {
+        guard let attributes = attributes else {
+            return
+        }
+        let rawString = self.string
+        let scanner = Scanner(string: rawString)
+        let quotes = scanner.scanQuotesText()
+        quotes.forEach {
+            if let itemRange = rawString.range(of: $0) {
+                let range = NSRange(itemRange, in: rawString)
+                self.addAttributes(attributes, range: range)
+            }
+        }
+    }
+}
