+++ WordPressKit/NSMutableParagraphStyle+extensions.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+extension NSMutableParagraphStyle {
+    @objc convenience init(minLineHeight: CGFloat, lineBreakMode: NSLineBreakMode, alignment: NSTextAlignment) {
+        self.init()
+        self.minimumLineHeight  = minLineHeight
+        self.lineBreakMode      = lineBreakMode
+        self.alignment          = alignment
+    }
+
+    @objc convenience init(minLineHeight: CGFloat, maxLineHeight: CGFloat, lineBreakMode: NSLineBreakMode, alignment: NSTextAlignment) {
+        self.init(minLineHeight: minLineHeight, lineBreakMode: lineBreakMode, alignment: alignment)
+        self.maximumLineHeight  = maxLineHeight
+    }
+}
